refactor: Update zsh configuration and install dependencies
This change refactors the `.zshrc` file to update the aliases for
showing/hiding hidden files in Finder. Additionally, the `kubens` alias
is removed and replaced with `kns`. Mobile development tools are now
enabled instead of Flutter tools. In the `bootstrap.zsh` file, the
dependency installations for generic tools are extended to include the
`watch` command. Moreover, the `bdw-gc` package is now installed for C++
tools, and the `dotnet` package is installed for Unity tools. These
changes aim to improve the shell experience, provide enhanced tools for
different development contexts, and ensure better compatibility with
Mobile and Unity development. No issues are referenced.